Support both Docker Compose v1 and v2
I don't want to drop v1 support yet since so many folks are still using
it and it's technically still opt-in for Docker Desktop at the time of
making this commit.

Eventually this project will always use v2 but for now this seems like a
reasonable middleground to support both. The only real cost is if you
customize the run script with Docker Compose commands you need to
remember to reference $docker_compose instead of calling
`docker-compose` or `docker compose` directly.

If anyone is curious I've fully moved to using Docker Compose v2 in
development and production.